Move repo field to shared insight form components
