+++ client/web/src/insights/components/form/repositories-field/components/flex-textarea/FlexTextarea.module.scss
@@ -1,5 +1,5 @@
 .textarea {
-    margin: 0; // todo: [crafting-normalize] reset textarea margin in FF
+    margin: 0;
     overflow: auto; // for ie, https://css-tricks.com/textarea-tricks/#article-header-id-8
     resize: none;
     box-sizing: border-box;
