--- client/web/src/insights/pages/creation/search-insight/components/repositories-field/RepositoriesField.module.scss
@@ -1,9 +1,0 @@
-.combobox {
-    width: 100%;
-
-    &-popover {
-        border: solid 1px var(--border-color);
-        box-shadow: var(--box-shadow);
-        background-color: var(--input-bg);
-    }
-}
