--- client/web/src/insights/pages/creation/search-insight/components/repositories-field/components/suggestion-panel/SuggestionPanel.module.scss
@@ -1,28 +1,0 @@
-@import '@reach/combobox/styles.css';
-
-.loading-panel {
-    height: 100px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.suggestions-list {
-
-    &__item {
-        padding: 0.5rem;
-
-        &:hover {
-            background-color: var(--body-bg);
-        }
-    }
-
-    [data-suggested-value] {
-        font-weight: normal;
-    }
-
-    [data-user-value] {
-        font-weight: bold;
-    }
-}
-
